Flatten send/uni.rs into send/mod.rs
